+++ CHANGELOG.md
@@ -6,8 +6,10 @@
     * Added `Grav\Framework\ContentBlock` classes for nested HTML blocks with CSS/JS assets
     * Added `Grav\Framework\Object` classes for creating collections of objects
     * Added `Grav\Framework\Page` interfaces
-    * Deprecated GravTrait
+    * Added `$grav->login()` and `$grav->logout()` functions with event hooks for plugins
+    * Added events `onUserLoginAuthenticate`, `onUserLoginAuthorize`, `onUserLoginFailure`, `onUserLogin`, `onUserLogout`
     * Added `|nicenumber` Twig filter
+    * Deprecated GravTrait
 1. [](#improved)
     * Make it possible to include debug bar also into non-HTML responses